--- crawler_project/crawler_project/spiders/output.json
@@ -1,11 +1,0 @@
-[
-
-][
-
-][
-
-][
-
-][
-
-]
